feat: add toast feature

+++ app/(admin)/admin/SideMenuComponents/users.tsx
@@ -6,7 +6,8 @@ import { Loader2, ArrowLeft, ArrowRight, Edit, Trash } from "lucide-react";
 import EditUser from "../Components/editForm";
 import AddUserForm from "../Form/Form";
 import { useFormContext } from "../Context";
-
+import { toast } from "sonner";
+import { useAuthStore } from "@/app/(auth)/store/userAuth";
 interface userType {
   id: string;
   email: string;
@@ -28,6 +29,7 @@ interface GetUsersResponse {
 }
 
 export default function User() {
+  const { user: loggedInUser } = useAuthStore();
   const [page, setPage] = useState(1);
   const {
     isAddUserFormOpen,
@@ -99,6 +101,10 @@ export default function User() {
   }
 
   async function confirmDelete(userId: string) {
+    if (userId === loggedInUser?.id) {
+      toast.error("You cannot delete your own account!");
+      return;
+    }
     const res = await fetch("/api/delete", {
       method: "DELETE",
       headers: {
@@ -108,8 +114,8 @@ export default function User() {
     });
     const deleteData = await res.json();
     if (deleteData.success) {
-      alert("User deleted successfully");
       refetch();
+      toast.success("User deleted successfully!!");
     } else {
       alert("Error deleting user");
     }
